Engine logic split from command
The /engine command has a lot of logic and was rather large to be contained within the command function. So now the core logic of the command is in its own function called "StartVehicleAttempt".

The SpawnTimer function now handles setting the player skin (in case it doesn't set the first time), and the player's gravity depending on if they have unlocked a certain zombie perk or not.